Demultiplex fastq file pairs and trim tags and primers (reverse strand not checked)

Same as demultiplex_and_trim_fastq, but without checking the reverse-complement of the sequences.

FASTQ file pairs are first demultiplexed by requiring a perfect match between the tag sequence and the 5' end of the read. The resulting files are then trimmed to remove primer sequences using less stringent matching parameters (controlled by cutadapt_error_rate). A match between the primer and the 5' end of the read is required for the read to be trimmed and retained. Matching of the 3' primer is optional

Input files can be compressed or uncompressed. Output compression is controlled by compress.

Usage

demultiplex_fastq_pairs_strand_plus(
  fastqinfo,
  fastq_dir,
  outdir,
  cutadapt_path = NULL,
  num_threads = 0,
  tag_to_end = T,
  primer_to_end = T,
  cutadapt_error_rate = 0.1,
  sep = ",",
  compress = F,
  quiet = T
)

Arguments

fastqinfo

Data frame or path to a CSV file with the following columns: tag_fw, primer_fw, tag_rv, primer_rv, sample, sample_type (mock/negative/real), habitat (optional), replicate, fastq_fw, fastq_rv

fastq_dir

Character string specifying the directory containing input FASTQ files (listed in the fastq_fw and fastq_rv columns of fastqinfo).

outdir

Character string specifying the output directory.

cutadapt_path

Optional path to the cutadapt executable. If NULL (default), resolved via getOption("vtamR.cutadapt_path") or the PATH (see get_path).

num_threads

Positive integer specifying the number of CPU threads to use. If 0, all available CPUs are used.

tag_to_end

Logical. If TRUE, tags are assumed to be located at the extremities of reads (starting at the first base).

primer_to_end

Logical. If TRUE, primers are assumed to follow directly after tags (i.e., no heterogeneity spacer).

cutadapt_error_rate

Numeric value between 0 and 1 specifying the maximum allowed error rate between primers and reads (exact match is required for tags).

sep

Character string specifying the field separator used in input and output CSV files.

compress

Logical. If TRUE, compress output files using gzip.

quiet

Logical. If TRUE, suppress informational messages and only display warnings or errors.

Value

Data frame similar to the input fastqinfo file, but contains the output fastq file names and read counts.

Examples

if (FALSE) { # \dontrun{
fastqinfo_df <- demultiplex_fastq_pairs_strand_plus(
  fastqinfo = fastqinfo_df,
  fastq_dir = "data/fastq",
  outdir = "data/fastq_demultiplexed",
  tag_to_end = TRUE,
  primer_to_end = TRUE,
  sep = ","
)
} # }