Download from OSF

Downloads a file from the Open Science Framework (OSF). If the file is a .tar.gz archive, it can optionally be extracted after download.

Usage

download_osf(filename, url, dest_dir = ".", untar = TRUE, quiet = FALSE)

Arguments

filename

Character string specifying the name of the file to download.

url

Character string; direct download URL of the OSF file.

dest_dir

Character string; path to the directory where the file should be downloaded.

untar

Logical; if TRUE, the archive is extracted after download.

quiet

Logical; if TRUE, suppress informational messages and show only warnings or errors.

Value

Invisible full path to the directory where the file was downloaded.

Examples

if (FALSE) { # \dontrun{
download_osf(
  filename = "COInr_for_vtam_2025_05_23_dbV5.tar.gz",
  url = "https://osf.io/download/jyhz6/",
  dest_dir = "~/vtamR/OSF",
  untar = TRUE,
  quiet = FALSE
)
} # }