Determine the Lowest Taxonomic Group (LTG)

Identify the Lowest Taxonomic Group (LTG) that contains at least phit percent of the input taxids.

Usage

find_ltg(taxids, lineages, phit = 70)

Arguments

taxids

Vector of taxids (taxonomic identifiers), possibly containing duplicates.

lineages

Data frame with taxids in the first column followed by their taxonomic lineages (represented as taxids, starting from the lowest resolution).

phit

Integer between 0 and 100 specifying the minimum percentage of taxids that must be included in the LTG.

Value

Numeric taxid corresponding to the LTG, or NA if no LTG can be determined.

Examples