Get ranked taxonomic lineages

Retrieve major taxonomic ranks for each input taxid based on a reference taxonomy table.

Usage

get_ranked_lineages(taxids, tax_df, fill_lineage = TRUE)

Arguments

taxids

Vector of taxids (taxonomic identifiers).

tax_df

Data frame with the following columns: tax_id, parent_tax_id, rank, name_txt, taxlevel (8: species, 7: genus, 6: family, 5: order, 4: class, 3: phylum, 2: kingdom, 1: domain, 0: root).

fill_lineage

Logical. If TRUE, fill missing higher-level taxa using the name of the next known lower-level taxon, prefixed by the corresponding rank (e.g., No_kingdom_Chrysophyceae if kingdom is missing but class is known).

Value

Data frame containing ranked lineages with columns: ltg_taxid, ltg_name, ltg_rank, ltg_rank_index, domain_taxid, domain, kingdom_taxid, kingdom, phylum_taxid, phylum, class_taxid, class, order_taxid, order, family_taxid, family, genus_taxid, genus, species_taxid, species.

Examples

if (FALSE) { # \dontrun{
taxids <- c(9593, 9606)
tax_df <- read.delim("xxxxxxxx", header = TRUE, sep = "\t", fill = TRUE, quote = "")
get_ranked_lineages(taxids, tax_df)
} # }