get_stat

Compute summary statistics (reads, variants, samples, and replicates) and append them to a statistics data frame.

Usage

get_stat(
  read_count,
  stat_df = NULL,
  stage = "",
  params = NA,
  outfile = NULL,
  sep = ","
)

Arguments

read_count

Data frame or path to a CSV file containing the following variables: asv_id, sample``, replicate, read_count, asv.

stat_df

A data frame containing the following variables:

`parameters` Parameters used for the analysis
`asv_count` Number of ASVs detected
`read_count` Total number of reads
`sample_count` Number of samples analyzed
`sample_replicate_count` Number of replicates per sample

If provided, the function appends a new row to stat_df. If not provided, a new data frame with a single row is initialized and returned.

stage

Character string specifying the name of the filtering step. It is used as the row name in stat_df.

params

Character string containing concatenated parameter values used for the filtering step.

outfile

Character string specifying the name of a CSV file to write the updated data frame. If NULL, no file is written.

sep

Character string specifying the field separator used in input and output CSV files.

Value

Data frame. The updated stat_df with an additional row.

Examples

if (FALSE) { # \dontrun{
get_stat(
  read_count_df,
  stat_df,
  stage = "filter_occurrence_variant",
  params = "0.002;by_replicate=TRUE"
)

get_stat(
  read_count_df,
  stat_df = NULL,
  stage = "filter_indel",
  params = "0.002;by_replicate=TRUE",
  outfile = "out/ReadCount_stat.csv"
)
} # }